[#141] chg: form structure
